Handle git worktrees
In a git worktree, `.git` is not a directory. Because of Windows
contaminating Linux, it's not even a symlink, it's a file with a path
stored in it. Handle that by no longer referring to `.git/big`.

Put the big dir in $GIT_COMMON_DIR so it's shared between worktrees.